ai-service/requirements.txt
flask==2.3.3
flask-cors==4.0.0
opencv-python==4.8.1.78
pillow==10.0.1
numpy==1.24.3
easyocr==1.7.0
googletrans==4.0.0rc1
requests==2.31.0
torch==2.0.1
torchvision==0.15.2
diffusers==0.21.4
transformers==4.33.2
accelerate==0.23.0